third_party/WebKit/LayoutTests/http/tests/htmlimports/cross-origin.html

<!DOCTYPE html>
<html>
<head>
<link id="target" rel="import" href="http://localhost:8080/htmlimports/resources/hello.html">
<script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<script>
shouldBeNull("target.import");
</script>
</body>
</html>
